FIX: separate getTimeAgo.js, formatDate.js
